python: fix Xcore bug

+++ bindings/python/capstone/xcore.py
@@ -15,7 +15,7 @@ class XcoreOpMem(ctypes.Structure):
 class XcoreOpValue(ctypes.Union):
     _fields_ = (
         ('reg', ctypes.c_uint),
-        ('imm', ctypes.c_int64),
+        ('imm', ctypes.c_int32),
         ('mem', XcoreOpMem),
     )
